insert record for dr

+++ web_module/trunk/src/main/com/topcoder/apps/review/projecttracker/ProjectUtil.java
@@ -508,7 +508,7 @@ public class ProjectUtil {
     }
 
     private static void prepareProjectInfo(Connection conn, long compVersId, long projectId, long modUserId,
-                                           long forumCategoryId, final double price, final int projectTypeId) throws SQLException {
+                                           long forumCategoryId, final double price, final long projectTypeId) throws SQLException {
         System.out.println("price " + price);
         PreparedStatement ps = null;
         ResultSet rs = null;
